1508 init overlay
