+++ web-client/src/views/FileDocument/FileDocumentWizard.jsx
@@ -7,7 +7,6 @@ import { SelectDocumentType } from './SelectDocumentType';
 import { SuccessNotification } from '../SuccessNotification';
 import { Tab, Tabs } from '../../ustc-ui/Tabs/Tabs';
 import { ViewAllDocuments } from './ViewAllDocuments';
-import { ViewDocumentCategory } from './ViewDocumentCategory';
 import { connect } from '@cerebral/react';
 import { state } from 'cerebral';
 import React from 'react';
