+++ web-client/src/views/FileDocument/ViewAllDocuments.jsx
@@ -1,6 +1,4 @@
-import { Mobile, NonMobile } from '../../ustc-ui/Responsive/Responsive';
 import { ViewAllDocumentsDesktop } from './ViewAllDocumentsDesktop';
-import { ViewAllDocumentsMobile } from './ViewAllDocumentsMobile';
 import { connect } from '@cerebral/react';
 import React from 'react';
 
@@ -9,12 +7,7 @@ export const ViewAllDocuments = connect(
   () => {
     return (
       <React.Fragment>
-        <NonMobile>
-          <ViewAllDocumentsDesktop />
-        </NonMobile>
-        <Mobile>
-          <ViewAllDocumentsMobile />
-        </Mobile>
+        <ViewAllDocumentsDesktop />
       </React.Fragment>
     );
   },
